Move kube_job functions into a module
